requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via
    #   tensorboard
    #   tensorflow
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via starlette
astunparse==1.6.3
    # via tensorflow
cachetools==5.5.0
    # via google-auth
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via click
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.6
    # via -r requirements.in
filelock==3.16.1
    # via torch
flatbuffers==24.12.23
    # via tensorflow
fonttools==4.55.3
    # via matplotlib
fsspec==2024.12.0
    # via torch
gast==0.4.0
    # via tensorflow
google-auth==2.37.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.68.1
    # via
    #   tensorboard
    #   tensorflow
h11==0.14.0
    # via uvicorn
h5py==3.12.1
    # via tensorflow
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.5
    # via torch
keras==2.10.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
kiwisolver==1.4.7
    # via matplotlib
libclang==18.1.1
    # via tensorflow
markdown==3.7
    # via tensorboard
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.9.4
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   h5py
    #   keras-preprocessing
    #   matplotlib
    #   opencv-python
    #   tensorboard
    #   tensorflow
    #   torchvision
oauthlib==3.2.2
    # via requests-oauthlib
opencv-python==4.10.0.84
    # via -r requirements.in
opt-einsum==3.4.0
    # via tensorflow
packaging==24.2
    # via
    #   matplotlib
    #   tensorflow
pillow==11.1.0
    # via
    #   matplotlib
    #   torchvision
protobuf==3.19.6
    # via
    #   tensorboard
    #   tensorflow
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.4
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
python-dotenv==1.0.1
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
requests==2.32.3
    # via
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   keras-preprocessing
    #   python-dateutil
    #   tensorflow
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
sympy==1.13.1
    # via torch
tensorboard==2.10.1
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.10.1
    # via -r requirements.in
tensorflow-estimator==2.10.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.31.0
    # via tensorflow
termcolor==2.5.0
    # via tensorflow
torch==1.12.1+cu113
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.13.1+cu113
    # via -r requirements.in
-f https://download.pytorch.org/whl/torch_stable.html
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   tensorflow
    #   torch
    #   uvicorn
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
werkzeug==3.1.3
    # via tensorboard
wheel==0.45.1
    # via
    #   astunparse
    #   tensorboard
wrapt==1.17.0
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
